docs: remove decorative table row from README header

+++ frontend/src/hooks/useAttendantPushNotifications.ts
@@ -79,7 +79,7 @@ export function useAttendantPushNotifications(token: string): UseAttendantPushNo
       const reg = await navigator.serviceWorker.ready;
       const sub = await reg.pushManager.subscribe({
         userVisibleOnly:      true,
-        applicationServerKey: urlBase64ToUint8Array(publicKey),
+        applicationServerKey: urlBase64ToUint8Array(publicKey) as Uint8Array,
       });
 
       // 4. Extract keys as URL-safe base64
